refactor utils.js
